move EnvVar to cli_config
